Moved TileIndices to its own file
