+++ MegaManLofi/TileIndices.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace MegaManLofi
+{
+   struct TileIndices
+   {
+      long long Left = 0;
+      long long Top = 0;
+      long long Right = 0;
+      long long Bottom = 0;
+   };
+}
